requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
xendit-python==7.0.0
